song_mashup/models/audio_model.py
"""
AI models for audio analysis and generation.
Implements wrappers for MusicBERT and Suno Bark models.
"""

import os
import numpy as np
import torch
from typing import Dict, List, Optional, Tuple, Union
import librosa

# Import these conditionally to avoid errors if not installed
try:
    from transformers import AutoModel, AutoTokenizer
    TRANSFORMERS_AVAILABLE = True
except ImportError:
    TRANSFORMERS_AVAILABLE = False

# Updated Bark import for newer version
try:
    from bark import generate_audio, SAMPLE_RATE
    BARK_AVAILABLE = True
except ImportError:
    try:
        # Fallback for newer versions of suno-bark package
        from bark import generation
        SAMPLE_RATE = generation.SAMPLE_RATE
        BARK_AVAILABLE = True
    except ImportError:
        BARK_AVAILABLE = False


class MusicBERTAnalyzer:
    """
    Wrapper for MusicBERT model to analyze musical features.
    """
    
    def __init__(self, model_path: Optional[str] = None):
        """
        Initialize the MusicBERT analyzer.
        
        Args:
            model_path: Path to a pre-trained MusicBERT model, or None to use default
        """
        if not TRANSFORMERS_AVAILABLE:
            raise ImportError("Transformers library is required for MusicBERT")
        
        self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        
        # Use the default model if none specified
        if model_path is None:
            model_path = "microsoft/musicbert-base"
        
        # Load model and tokenizer
        try:
            self.tokenizer = AutoTokenizer.from_pretrained(model_path)
            self.model = AutoModel.from_pretrained(model_path).to(self.device)
            self.model.eval()  # Set to evaluation mode
        except Exception as e:
            raise RuntimeError(f"Failed to load MusicBERT model: {str(e)}")
    
    def analyze(self, audio_features: Dict) -> Dict:
        """
        Analyze audio features using MusicBERT.
        
        Args:
            audio_features: Dictionary containing audio features
            
        Returns:
            Dictionary with analysis results
        """
        # This is a placeholder for actual MusicBERT analysis
        # In a real implementation, we would:
        # 1. Convert audio features to MIDI-like tokens or other representation
        # 2. Feed these to the MusicBERT model
        # 3. Extract embeddings or predictions
        
        # Simulate analysis with random values
        analysis = {
            'embeddings': np.random.randn(768),  # Simulated BERT embeddings
            'melody_score': float(np.random.uniform(0.5, 0.9)),
            'harmony_score': float(np.random.uniform(0.5, 0.9)),
            'rhythm_score': float(np.random.uniform(0.5, 0.9))
        }
        
        return analysis


class BarkGenerator:
    """
    Wrapper for Suno Bark model to generate audio.
    Updated to work with newer versions of Bark.
    """
    
    def __init__(self):
        """Initialize the Bark generator."""
        if not BARK_AVAILABLE:
            raise ImportError("Suno Bark library is required")
        
        # Check if we're using the newer API
        self.has_new_api = hasattr(generation, "preload_models") if "generation" in globals() else False
        
        # Preload models if available in this version
        try:
            if self.has_new_api:
                generation.preload_models()
            else:
                # Older versions might have different initialization
                pass
                
            self.device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
        except Exception as e:
            raise RuntimeError(f"Failed to initialize Bark model: {str(e)}")
    
    def generate(self, prompt: str, duration_sec: float = 10.0) -> np.ndarray:
        """
        Generate audio based on text prompt.
        
        Args:
            prompt: Text description of the audio to generate
            duration_sec: Duration of audio to generate in seconds
            
        Returns:
            NumPy array containing generated audio
        """
        try:
            # Attempt to use Bark to generate audio based on the prompt
            if self.has_new_api:
                # New API (0.1.5+)
                audio_array = generation.generate_audio(prompt)
            else:
                # Fall back to older API if available
                audio_array = generate_audio(prompt)
                
            return audio_array
            
        except Exception as e:
            print(f"Error generating audio with Bark: {str(e)}")
            
            # Fall back to a simple tone generation if Bark fails
            sample_rate = SAMPLE_RATE
            t = np.linspace(0, duration_sec, int(duration_sec * sample_rate))
            
            # Generate a simple chord
            tone = (
                0.5 * np.sin(2 * np.pi * 440 * t) +  # A4
                0.3 * np.sin(2 * np.pi * 554 * t) +  # C#5
                0.2 * np.sin(2 * np.pi * 659 * t)    # E5
            )
            
            return tone


def blend_with_ai(audio_files: List[Dict], blend_factors: List[float]) -> np.ndarray:
    """
    Use AI to intelligently blend multiple audio files.
    
    Args:
        audio_files: List of dictionaries containing audio data
        blend_factors: List of blend factors for each file
        
    Returns:
        NumPy array containing blended audio
    """
    # This would implement a more sophisticated blending algorithm
    # using the AI models to guide the process
    
    # For now, return a simple blend
    sample_rate = audio_files[0]['sample_rate']
    min_length = min(len(audio['waveform']) for audio in audio_files)
    
    # Initialize output array
    output = np.zeros(min_length)
    
    # Simple weighted sum
    for audio, factor in zip(audio_files, blend_factors):
        # Truncate to min length
        waveform = audio['waveform'][:min_length]
        # Add weighted contribution
        output += waveform * factor
    
    # Normalize
    output = librosa.util.normalize(output)
    
    return output 