requirements.txt
# Core data processing
numpy==1.24.3
scipy==1.10.1
pandas==1.5.3

# Audio processing
librosa==0.10.1
pydub==0.25.1
soundfile==0.12.1
ffmpeg-python==0.2.0
demucs==4.0.0

# Machine Learning
torch==2.0.1
torchaudio==2.0.2
transformers==4.30.2
scikit-learn==1.3.0
matplotlib==3.7.2

# AI Audio Models
suno-bark==0.1.5
# audiocraft requires Python 3.9 or less; commenting out for now
# audiocraft==0.0.1

# Web UI
streamlit==1.24.0
python-multipart==0.0.6

# For potential background processing
joblib==1.2.0

# Additional dependencies for enhanced AI models
scikit-learn>=1.3.0
librosa>=0.10.1
resampy>=0.4.2
matplotlib>=3.7.2
seaborn>=0.12.2